feat: Include error code frame in jest results, not in console log
This is achieved with a custom jest matcher.

The benefit is that the error is now grouped with the test, so that
now when multiple tests are run all at once, this error code frame
isn't sitting at the top of all the tests in the "console log
section", instead it's immediately underneath the description of
the test.