feat: Allow for standard test() & it() in type def tests
Rather than looking for a single line comment with "// test:", just
change the regexes up a bit to allow people to use the test('
description, () => {}) syntax that they're already familiar with.

Because the .test-d.ts file is compiled and not run, it makes no
difference if the tests (whether they be expectX() calls or just
a line of typescript that should not throw an error) is inside
of some arbitrary function (i.e. test()) or at the root of the
file.