feat: Read and write files to allow editing the compiled file contents
I wish this wasn't necessary because obviously file reading and
writing is kind of slow.

But so far as I can tell, both tsd-lite and @tsd/typescript only
accept a file path, not a string for the file, so I don't have a
way to manipulate the file to support test.skip() or test.only()
without writing to a separate file.

---

NOTE!!!

The node-tmp and tmp-promise libraries both say repeatedly that
it is not necessary to manually call the cleanup() function. I'm
going to trust that in that because awiating the cleanup function
really slows down the tests.

See:
https://github.com/raszi/node-tmp#asynchronous-file-creation
https://github.com/benjamingr/tmp-promise#asynchronous-file-creation