test(get-tsd-results): Write tests
Create actual type definition tests that need to get compiled.

One of the tests is failing because when I copy fhe file to the test
directory, it doesn't have acess to the imports it needs.

I think the only way I can fix that is by creating the tmp file
in the same directory as the original test file.