fix: Write tmp file to same dir so that imports resolve properly
I was trying originally to write this file to a TMP directory so that
the user would never be able to see that it exists in its short
lifespan, but that created the problem that any imports this file has
would no longer resolve properly.

So the only solution to that without a ton of work to manipulate the
import paths (those in node_modules would be particularly tricky
because you would have to read the package.json of every one) is to
create this tmp file in the same directory as the test, and then
immediately delete it of course.

The user might see this file and think it's kind of weird, but they
might not notice it at all.