fix(comment-out-only): Index was relative to substring
In the getIndicesOfAllOnlys() function, I was going through the string
and looking for instances of only functions. Once I found one, I was
adding that index to the array, and then for the next iteration
creating a substring that excluded the only that was just found so
that we could find the next one.

The problem was that the index that was being added to the array to be
returned was relative to the substring, instead of being relative to
the entire file. So of course when those indices were interpreted as
being relative to the whole string, they were incorrect.